requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
fastapi==0.115.5
uvicorn==0.32.1
pydantic>=2.11.7
python-dotenv>=1.1.0
mcp>=1.17.0
playwright==1.49.0
